requirements.txt
streamlit
yfinance
joblib
scikit-learn
pandas
numpy
kiteconnect
plotly
websocket-client